+++ app/src/main/java/woowacourse/movie/feature/reservation/reserve/MovieDetailActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.feature.reservation
+package woowacourse.movie.feature.reservation.reserve
 
 import android.content.Context
 import android.content.Intent
@@ -9,6 +9,8 @@ import com.example.domain.usecase.GetMovieRunningTimeUseCase
 import woowacourse.movie.R
 import woowacourse.movie.feature.BackKeyActionBarActivity
 import woowacourse.movie.feature.main.MainActivity.Companion.KEY_MOVIE
+import woowacourse.movie.feature.reservation.reserve.selection.DateTimeSpinner
+import woowacourse.movie.feature.reservation.reserve.selection.ReservationCounter
 import woowacourse.movie.feature.seat.SeatSelectActivity
 import woowacourse.movie.model.CountState
 import woowacourse.movie.model.MovieState
