+++ app/src/main/java/woowacourse/movie/feature/reservation/reserve/selection/DateTimeSpinner.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.feature.reservation
+package woowacourse.movie.feature.reservation.reserve.selection
 
 import android.view.View
 import android.widget.Spinner
