+++ app/src/main/java/woowacourse/movie/feature/reservation/reserve/selection/ReservationCounter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.feature.reservation
+package woowacourse.movie.feature.reservation.reserve.selection
 
 import android.view.View
 import android.widget.Button
