+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.reservation.MovieDetailActivity">
+    tools:context=".feature.reservation.reserve.MovieDetailActivity">
 
     <ScrollView
         android:layout_width="match_parent"
